implement payment resource

+++ app/graphql/mutations/remove_product_from_cart.rb
@@ -3,14 +3,25 @@
 module Mutations
   class RemoveProductFromCart < BaseMutation
     argument :product_detail_id, ID, required: true
-    field :cart, Types::CartType, null: true
+    argument :all, Boolean, required: false
+    field :product_removed, Types::CartType, null: true
+    field :message, String, null: true
 
-    def resolve(product_detail_id:)
+    def resolve(product_detail_id:, all: false)
       authorize_user
       cart = Cart.find_by!(product_detail_id: product_detail_id, user_id: context[:current_user].id)
-      cart.destroy
+      if all
+        cart.destroy
+        return {
+          product_removed: cart,
+          message: "Products successfully removed from cart"
+        }
+      end
+      product_removed = cart.decrement!(:quantity)
+      product_removed.destroy if product_removed.quantity <= 0
       {
-        cart: cart
+        product_removed: product_removed,
+        message: "Product successfully removed from cart"
       }
     end
   end
